Delete dead code: pure-forwarder dispatch_input, painting.cpp stub, unused makers
- Remove dispatch_input wrapper in input.hpp; window.cpp now calls
  dispatch_keyboard and dispatch_mouse directly. Saves one indirection layer.
- Collapse input_keyboard.hpp, input_mouse.hpp, input_core.hpp into a
  single input.hpp. They each held one declaration with no other content.
- Fold painting.cpp (28 LOC) into painting_scene.cpp and delete
  painting.hpp; paint_all is now declared next to paint_scene where it
  belongs.
- Drop make_scrollable_view and make_dropdown free-function wrappers in
  ui_style.hpp; both had zero call sites.

Verified: clang++ build + 247/247 tests pass on Linux.
