--- src/input_core.hpp
@@ -1,5 +1,0 @@
-#pragma once
-#include <windows.h>
-#include "wndstate.hpp"
-
-void handle(HWND hwnd, int act, WndState& s);
