--- src/input_mouse.hpp
@@ -1,6 +1,0 @@
-#pragma once
-#include <windows.h>
-#include <optional>
-#include "wndstate.hpp"
-
-std::optional<LRESULT> dispatch_mouse(HWND hwnd, UINT msg, WPARAM wp, LPARAM lp, WndState& s);
