--- src/painting.hpp
@@ -1,5 +1,0 @@
-#pragma once
-#include <windows.h>
-#include "paint_ctx.hpp"
-
-void paint_all(HDC hdc, int cw, int ch, PaintCtx& ctx);
