+++ src/painting_scene.hpp
@@ -10,3 +10,6 @@
 // Text Ops currently render with the toolbar font (fontSm). A future pass will
 // extend ui_scene::Op with a font/style hint as more widgets migrate.
 void paint_scene(HDC hdc, const ui_scene::Scene& scene, PaintCtx& ctx);
+
+// Top-level paint entry: builds the main scene from app state and renders it.
+void paint_all(HDC hdc, int cw, int ch, PaintCtx& ctx);
